Add files via upload
This is a neural netowrk that consist of two layers (h1 and h2) with one out output (o1). With two inputs, there are six weights and three biases. The netowrk has a learning rate of 0.01 and uses a sigmoid activation function and mean squared loss. Stochastic gradient descent (SGD) was used to train the model. Backpropagation was used to calculate the partial derivatives of the loss.